[libc][stdio] Add buffering for fread and fwrite

+++ libs/libc/include/unistd.h
@@ -13,8 +13,8 @@ int execve(const char* path, char** argv, char** env);
 int wait(int pid);
 
 /* fs */
-int read(int fd, char* buf, size_t count);
-int write(int fd, const void* buf, size_t count);
+ssize_t read(int fd, char* buf, size_t count);
+ssize_t write(int fd, const void* buf, size_t count);
 
 __END_DECLS
